mma
git-svn-id: http://10.1.8.14:18080/svn/Local_Business/DAT-Event/Club-Event@3533 5e73b0b3-88fa-41e2-a44a-e6aac4854029

+++ src/main/java/com/dat/event/dto/SupportedMember.java
@@ -1,4 +1,4 @@
 package com.dat.event.dto;
 
-public record SupportedMember(int memberId, String name, String month) {
+public record SupportedMember(Long memberId, String name, String month) {
 }
